+++ src/client/components/Inputs/Select/types.ts
@@ -1,3 +1,5 @@
 import { Label } from 'meta/assessment'
 
 export type Option = { label: Label; value: string }
+
+export type TranslatedOption = { label: string; value: string }
